+++ RELEASE-NOTES.md
@@ -13,6 +13,9 @@ as release notes for end user on a Recalbox upgrade.
 - Add support for Anbernic RG351p and RG351m (Thanks lxmaster1 & yomax69fly !)
 - Add support for Realtek RTW-88 series of wifi controllers
 - Add support for N64 and Megadrive/Genesis controllers from Nintendo Switch Online
+- Add support for HD mode on Pi5, PC (Dreamcast, Naomi, Atomiswave, Playstation, Saturn)
+- Add support for Widescreen mode on Pi4, Pi5, PC (Dreamcast, Naomi, Atomiswave, Snes, Megadrive)
+
 
 ### News
 - Add Arcade View